AI/requirements.txt
# -------------------------------
# Web Backend
# -------------------------------
fastapi
uvicorn

# -------------------------------
# Environment & Config
# -------------------------------
python-dotenv

# -------------------------------
# LLM (Groq)
# -------------------------------
groq

# -------------------------------
# RAG & Embeddings
# -------------------------------
sentence-transformers
faiss-cpu
numpy

# -------------------------------
# PDF ingestion (RAG)
# -------------------------------
pypdf

# -------------------------------
# HTTP / External APIs
# -------------------------------
requests

# -------------------------------
# Resilience & Retries
tenacity

# -------------------------------
# Testing (dev)
pytest
pytest-asyncio
httpx
requests-mock

